implement manageprofile. many css changes.
