+++ frontend/src/components/ManageProfile/BlockedList.tsx
@@ -6,7 +6,8 @@ import { unblockUser } from "src/ApiCalls/userActions";
 import { IdAndNameDTO } from "user-dto";
 
 // CSS
-import "src/styles/modals.css";
+import "src/styles/style.css";
+import "src/styles/manageProfile.css";
 
 interface blockedListProps {
     id: number;
@@ -34,16 +35,16 @@ function BlockedList(props: blockedListProps) {
     return (
         <div>
             {!group || group.length === 0 ? (
-                <p>No toxic ppl... yet!</p>
+                <div className="p">No toxic ppl... yet!</div>
             ) : (
-                <table className="contacts-table">
-                    <tbody className="contacts-table">
+                <table className="modalUserList">
+                    <tbody>
                         {group.map((entry, index) => (
-                            <tr className="contacts-table" key={entry.id}>
-                                <td className="contacts-table"> {entry.name}</td>
-                                <td className="contacts-table">
+                            <tr key={entry.id}>
+                                <td> {entry.name}</td>
+                                <td>
                                     <button
-                                        className="contacts-button"
+                                        className="contactsButton"
                                         onClick={() => handleUnBlockUser(entry.id, index)}
                                     >
                                         ❌
