+++ frontend/src/components/ManageProfile/RequestOutList.tsx
@@ -6,7 +6,8 @@ import { cancelRequest } from "src/ApiCalls/userActions";
 import { IdAndNameDTO } from "user-dto";
 
 // CSS
-import "src/styles/modals.css";
+import "src/styles/style.css";
+import "src/styles/manageProfile.css";
 
 interface requestOutListProps {
     id: number;
@@ -34,16 +35,16 @@ function RequestOutList(props: requestOutListProps) {
     return (
         <div>
             {!group || group.length === 0 ? (
-                <p>No requests for now!</p>
+                <p>No outgoing requests. Go make some friends!</p>
             ) : (
-                <table className="contacts-table">
-                    <tbody className="contacts-table">
+                <table className="modalUserList">
+                    <tbody>
                         {group.map((entry, index) => (
-                            <tr className="contacts-table" key={entry.id}>
-                                <td className="contacts-table"> {entry.name}</td>
-                                <td className="contacts-table">
+                            <tr key={entry.id}>
+                                <td> {entry.name}</td>
+                                <td>
                                     <button
-                                        className="contacts-button"
+                                        className="contactsButton"
                                         onClick={() => handleCancel(entry.id, index)}
                                     >
                                         ❌
